rename file_service --> file_handling
